requirements.txt
psycopg2-binary
fastapi
uvicorn
strawberry-graphql
python-dotenv